bookStore

+++ controllers/CartController.js
@@ -1,8 +1,8 @@
+const ensureAuthorization = require("../auth");
 const conn = require("../mariadb");
 const {StatusCodes} = require("http-status-codes");
 const jwt = require("jsonwebtoken");
 const dotenv = require("dotenv");
-const res = require("express/lib/response");
 const {TokenExpiredError} = require("jsonwebtoken");
 dotenv.config();
 
@@ -17,7 +17,7 @@ const addToCart = (req, res) => {
         return res.status(StatusCodes.UNAUTHORIZED).json({
             'message': 'Token expired, try again later',
         });
-    } else if (authorization instanceof TokenExpiredError) {
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
 
         return res.status(StatusCodes.BAD_REQUEST).json({
             'message': 'Wrong token used, try again later',
@@ -48,7 +48,7 @@ const getCartItems = (req, res) => {
         return res.status(StatusCodes.UNAUTHORIZED).json({
             'message': 'Token expired, try again later',
         });
-    } else if (authorization instanceof TokenExpiredError) {
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
 
         return res.status(StatusCodes.BAD_REQUEST).json({
             'message': 'Wrong token used, try again later',
@@ -57,8 +57,13 @@ const getCartItems = (req, res) => {
         let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
-                WHERE user_id = ? AND cartItems.id IN (?)`;
+                WHERE user_id = ?`;
         let values = [authorization.id, selected];
+
+        if (selected) { // 주문서 작성 시 선택한 장바무니 목록 조회
+            sql += ` AND cartItems.id IN (?)`;
+            values.push(selected);
+        }
 
         conn.query(sql, values, (err, result) => {
             if (err) {
@@ -73,35 +78,30 @@ const getCartItems = (req, res) => {
 
 // 장바구니 삭제
 const removeFromCart = (req, res) => {
+    let authorization = ensureAuthorization(req, res);
 
-    const cartItemId = req.params.id; //cartItemId
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message': 'Token expired, try again later',
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
 
-    let sql = "DELETE FROM cartItems WHERE id = ?";
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            'message': 'Wrong token used, try again later',
+        })
+    } else {
+        const cartItemId = req.params.id; //cartItemId
 
-    conn.query(sql, cartItemId, (err, result) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+        let sql = "DELETE FROM cartItems WHERE id = ?";
 
-        return res.status(StatusCodes.OK).json(result);
-    })}
+        conn.query(sql, cartItemId, (err, result) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-function ensureAuthorization(req, res) {
-
-    try {
-        let receivedJWT = req.headers['authorization'];
-        console.log("receivedJWT : ", receivedJWT);
-
-        let decodedJWT = jwt.verify(receivedJWT, process.env.JWT_SECRET);
-        console.log(decodedJWT);
-
-        return decodedJWT;
-    } catch (err) {
-        console.log(err.name);
-        console.log(err.message);
-
-        return err;
+            return res.status(StatusCodes.OK).json(result);
+        })
     }
 }
