+++ controllers/OrderController.js
@@ -1,42 +1,57 @@
-// const conn = require("../mariadb");
+const ensureAuthorization = require("../auth");
+const jwt = require("jsonwebtoken");
 const {StatusCodes} = require('http-status-codes');
 const mariadb = require("mysql2/promise");
-const req = require("express/lib/request");
+const {TokenExpiredError} = require("jsonwebtoken");
 
 const order = async (req, res) => {
     const conn = await mariadb.createConnection({
         host: 'localhost', user: 'root', password: 'root', port: 3306, database: 'bookStore', dateStrings: true
     });
 
-    const {items, delivery, totalQuantity, totalPrice, userId, firstBookTitle} = req.body;
+    let authorization = ensureAuthorization(req, res);
 
-    // delivery 테이블 삽입
-    let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
-    let values = [delivery.address, delivery.receiver, delivery.contact];
-    let [result] = await conn.execute(sql, values);
-    let delivery_id = result.insertId;
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message': 'Token expired, try again later',
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
 
-    // orders 테이블 삽입
-    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) VALUES (?, ?, ?, ?, ?)`;
-    values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-    [result] = await conn.execute(sql, values);
-    let order_id = result.insertId;
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            'message': 'Wrong token used, try again later',
+        })
+    } else {
 
-    // items를 가지고, 장바구니에서 book_id와 quantity 조회
-    sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
-    let [orderItems, fields] = await conn.query(sql, [items]);
+        const {items, delivery, totalQuantity, totalPrice, firstBookTitle} = req.body;
 
-    // orderedBook 테이블 삽입
-    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
-    // items.. 배열 : 요소들을 하나씩 꺼내서 (foreach문 돌려서)
-    values = [];
-    orderItems.forEach((item) => {
-        values.push([order_id, item.book_id, item.quantity])
-    })
-    result = await conn.query(sql, [values]);
-    let results = await deleteCartItems(conn, items);
-    return res.status(StatusCodes.OK).json(results);
-};
+        // delivery 테이블 삽입
+        let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
+        let values = [delivery.address, delivery.receiver, delivery.contact];
+        let [result] = await conn.execute(sql, values);
+        let delivery_id = result.insertId;
+
+        // orders 테이블 삽입
+        sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) VALUES (?, ?, ?, ?, ?)`;
+        values = [firstBookTitle, totalQuantity, totalPrice, authorization.id, delivery_id];
+        [result] = await conn.execute(sql, values);
+        let order_id = result.insertId;
+
+        // items를 가지고, 장바구니에서 book_id와 quantity 조회
+        sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
+        let [orderItems, fields] = await conn.query(sql, [items]);
+
+        // orderedBook 테이블 삽입
+        sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
+        // items.. 배열 : 요소들을 하나씩 꺼내서 (foreach문 돌려서)
+        values = [];
+        orderItems.forEach((item) => {
+            values.push([order_id, item.book_id, item.quantity])
+        })
+        result = await conn.query(sql, [values]);
+        let results = await deleteCartItems(conn, items);
+        return res.status(StatusCodes.OK).json(results);
+    }
+}
 
 const deleteCartItems = async (conn, items) => {
     let sql = `DELETE FROM cartItems WHERE id IN (?)`;
@@ -55,21 +70,32 @@ const getOrders = async (req, res) => {
 
     let [rows, fields] = await conn.query(sql);
     return res.status(StatusCodes.OK).json(rows);
-
-
 };
 
 const getOrderDetails = async (req, res) => {
-    const {id} = req.params;
-    const conn = await mariadb.createConnection({
-        host: 'localhost', user: 'root', password: 'root', port: 3306, database: 'bookStore', dateStrings: true
-    });
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = `SELECT book_id, title, author, price, quantity FROM orderedBook LEFT JOIN books ON orderedBook.book_id = books.id WHERE order_id = ?`
-    let [rows, fields] = await conn.query(sql, [id]);
-    return res.status(StatusCodes.OK).json(rows);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message': 'Token expired, try again later',
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
 
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            'message': 'Wrong token used, try again later',
+        })
+    } else {
+        const orderId = req.params.id;
+        const conn = await mariadb.createConnection({
+            host: 'localhost', user: 'root', password: 'root', port: 3306, database: 'bookStore', dateStrings: true
+        });
+
+        let sql = `SELECT book_id, title, author, price, quantity FROM orderedBook LEFT JOIN books ON orderedBook.book_id = books.id WHERE order_id = ?`
+        let [rows, fields] = await conn.query(sql, [orderId]);
+        return res.status(StatusCodes.OK).json(rows);
+    }
 };
+
 
 module.exports = {
     order,
